remove unused method from client
